% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCSNBShiny.R
\name{KCSNBShiny}
\alias{KCSNBShiny}
\title{Naive Bayes Classifier}
\usage{
KCSNBShiny()
}
\description{
Predicts any variable in any categorical dataset for given values of predictor variables.If a dataset contains 4 variables, then any variable can be predicted based on the values of the other three variables given by the user. The user can upload their own datasets and select what variable they want to predict.A handsontable is provided to enter the predictor values and also accuracy of the prediction is also shown.
}
\examples{
if(interactive()){KCSNBShiny()}

}
